+++ db/migrate/20230331123537_add_state_to_offers.rb
@@ -1,0 +1,5 @@
+class AddStateToOffers < ActiveRecord::Migration[7.0]
+  def change
+    add_column :offers, :state, :integer, default: 0, null: false
+  end
+end