+++ test/models/offer_test.rb
@@ -1,0 +1,63 @@
+# frozen_string_literal: true
+
+require "test_helper"
+
+class OfferTest < ActiveSupport::TestCase
+  test "only one proposed or accepted offer is allowed for conversation at time" do
+    conversation = conversations(:one)
+    conversation.update!(offers: [offers(:one)])
+
+    assert_raises ActiveRecord::RecordInvalid do
+      Offer.create!(conversation:, start_date: Time.zone.today,
+        pay_rate_value: 50, pay_rate_time_unit: "hour", state: "proposed")
+    end
+
+    assert_raises ActiveRecord::RecordInvalid do
+      Offer.create!(conversation:, start_date: Time.zone.today,
+        pay_rate_value: 50, pay_rate_time_unit: "hour", state: "accepted")
+    end
+
+    assert_instance_of(
+      Offer,
+      Offer.create!(conversation:, start_date: Time.zone.today,
+        pay_rate_value: 50, pay_rate_time_unit: "hour", state: "declined")
+    )
+  end
+
+  test "#deleted_sender? returns true when there's no business assigned to offer's conversation" do
+    offer = offers(:one)
+
+    assert_not offer.deleted_sender?
+
+    offer.conversation.business.destroy
+    offer.reload
+
+    assert offer.deleted_sender?
+  end
+
+  test "#pay_rate_time_unit_key returns 'hour' when pay_rate_time_unit is 0" do
+    assert_equal Offer.new(pay_rate_time_unit: 0).pay_rate_time_unit_key, "hour"
+  end
+
+  test "#pay_rate_time_unit_key returns 'day' when pay_rate_time_unit is 1" do
+    assert_equal Offer.new(pay_rate_time_unit: 1).pay_rate_time_unit_key, "day"
+  end
+
+  test "#pay_rate_time_unit_key returns 'year' when pay_rate_time_unit is 2" do
+    assert_equal Offer.new(pay_rate_time_unit: 2).pay_rate_time_unit_key, "year"
+  end
+
+  test "#rate returns properly formatted string of pay_rate_value and pay_rate_time_unit" do
+    assert_equal(
+      Offer.new(pay_rate_value: 50, pay_rate_time_unit: Offer.pay_rate_time_units[:year]).rate,
+      "50.0/year"
+    )
+  end
+
+  def create_notification(message, recipient)
+    message.notifications_as_message.create!(recipient:,
+      type: NewMessageNotification.name,
+      params: {message:,
+               conversation: message.conversation})
+  end
+end